remove pk/hk from Update
